requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
Pillow==12.1.1
gunicorn==23.0.0
dj-database-url==2.2.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
